backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pydantic==2.5.3
python-dotenv==1.0.0
anthropic==0.18.1
docx2txt==0.8
pypdf==4.0.1
faiss-cpu==1.7.4
sentence-transformers==2.3.1
rapidfuzz==3.6.1
numpy==1.26.3
gunicorn==21.2.0
